requirements.txt
Django==4.2.7
Pillow==10.1.0
psycopg2-binary==2.9.9
django-crispy-forms==2.1
crispy-bootstrap5==0.7
whitenoise==6.6.0
gunicorn==21.2.0
python-decouple==3.8
dj-database-url==2.1.0
transbank-sdk==6.1.0
